requirements.txt
line-bot-sdk==1.3.0
urllib3==1.22
APScheduler==3.3.1
falcon==1.3.0
Cython==0.24
gevent==1.1.1
greenlet==0.4.9
gunicorn==19.4.5
python-mimeparse==1.5.1
requests==2.9.1
simplejson==3.8.2
six==1.10.0
requests-oauthlib==0.8.0
